akka-zeromq: Alter and access 'socketClosed' synchronously
Also, prior to closing the socket, unregister it from the poller. Note
that if the select recursion has terminated, queued requets from
'postStop' for closing the socket will not be processed.

Also, in the test specification, invoke ActorRef#stop and make the
'Close' request private to the 'akka.zeromq' package.

Signed-off-by: Karim Osman <karim@iki.fi>