akka-zeromq: Verify ZeroMQ installation and version in ConcurrentSocketActorSpec
If ZeroMQ is not installed, do not run any of the test cases in
ConcurrentSocketActorSpec. Instead, mark the tests pending and inform
about the missing installation. Previously, a too old version caused a
segfault when running the tests. Also, note that this requires version
0.0.3 of the ZeroMQ Scala Binding, and while at it, add ZeroMQ#version
for accessing the version of ZeroMQ.

Signed-off-by: Karim Osman <karim@iki.fi>