To avoid spurious errors in the Scala IDE plugin for Eclipse, the package declaration in any file should correspond to the phisical location where the file is placed.
This is a known limitation of the Eclipse platform (Java, the language, enforces this rule and Eclipse assumes it). We have an open ticket for this problem, but it is
not clear if we can do much about it, as it is an Eclipse design decision. http://scala-ide-portfolio.assembla.com/spaces/scala-ide/tickets/1000439

This change should have no consequences for library users, as the produced bytecode will be identical.