akka-zeromq: Add support for the rest of the socket configurations
While at it, add support for accessing values of 'ZMQ_FD' and
'ZMQ_RCVMORE' socket options. Note that accessing 'ZMQ_HWM' is not
currently supported as the implementation of the JZMQ like API is
missing an accessor method. This will be fixed in the next version of
the ZeroMQ Scala Binding. Also, remove some of the trailing whitespaces.

Signed-off-by: Karim Osman <karim@iki.fi>