Update to config lib 4bee83c
This is release 0.3.0 plus a patch to allow config.resource to start with '/'

The main change in this update is to fix more complex graphs of ${}
substitution references in the config file, some more complex usages
would confuse the library previously. This is a large patch which is
why it was not in 2.0.