Make provider registration of PromiseActorRefs lazy
By only registering a PromiseActorRef on access to `path`
we can shave off about 30% of the overhead involved in the common,
purely local `ask` use case.
Unfortunately `path` is accessed not only for ActorRef serialization
but also during `toString`, `equals` and `hashCode`, so one should
be careful how to handle PromiseActorRefs in hot paths.