Make method `dispatcher` in ActorRefFactory trait public (was protected)
ActorRefFactory is only extended by ActorSystem and ActorContext, both
of which make `dispatcher` public anyway. Writing code that requires
access to the dispatcher (e.g. for Promise creation) and is supposed
to work with ActorSystems as well as ActorContexts is easier, if the
ActorRefFactory allows public access to the dispatcher directly.