newActor(): replace 'emptyBehavior' to 'receive' after instance creation
Actor instances are allowed to call become()/unbecome() on
initialization. Given that 'inst.receive' isn't known until after
initialization, a placehold is put on the stack to represent it. In
the past become(discardOld=true) protected the last stack entry, and
the code to replace 'emptyBehavior' to 'receive' depended on that.

More simply (and works with new become()) is just to have code that
directly does the replace.