Adding the ActorDSL and the MigrationSystem.
The actorDSL provides a lightweight syntax to actor creation and unifies threads and actors in the similar fashion to original Scala actors. It exhibits very low performance so it should be used only for prototyping and experimentation. The idea behind the ActorDSL is to provide the similar functionality to Scala ActorDSL in Akka and thus ease the migration.

MigrationSystem is used for the migration of Scala Actors to the Akka based implementation. After the transition is complete it should be deprecated and removed from Akka.