=htc Parse Sunday dates in Set-Cookie headers
The Set-Cookie header would always fail to parse when given a an
expires date for a Sunday. I've updated the parsing code for both
"Sun" and "Sunday", however I have only managed to test "Sun". I
was never able to get an expires value for "Sunday", even though
the HTTP-date parsing code seems to suggest that long day names
are supported.