=mnt #17317 Harden reconnect when attaching to test conductor
The failing tests showed big variations in startup times, i.e.
JVM1 not started/bound when other JVMs tries to connect to the
test conductor. The reconnect logic was in place but it was
bound to 10 client-reconnects (10 seconds). It's better to let
it by default reconnect until the connect-timeout (20 seconds)
is elapsed.