=Fix regression of akka.util.ByteString#slice(...)
Fixes #21237

Currently slice(...) method unexpectedly throws an error for some arguments.
The issue comes from a corner-case of string _slice_ and we can fix this
by correctly handling 6 basic patterns(scenarios) described as follows:

Suppose the two input arguments are;
1) from, until <= 0
We check if `until <= 0` is met and return `ByteString.empty` if this is the case.

2) length <= from, until
We check if `length <= from` is met and return `ByteString.empty` if this is the case.

3) from <= 0 && length <= until
We return the byte string itself as we can _compact_ it.

4) 0 <= from <= until <= length
We _slice_ the byte string starting from `from` with the length of `until - from`

5) from < 0
Based on the condition 1, 2, 3 and 4, we know that there're two cases left,
i.e., either `from` or `until` is out of the _valid_ range, as we've checked
a) both of them are in the valid range in 4 and b) neither of them is in the valid range
(or we don't have to actually slice it where `from == 0` and `until == length`) in 3.
If `from` is the one that is out of the valid range then `0 <= until <= length` is satisfied,
which means we can return _slice_ from `0` to `until`.

6) length < until
Vice versa, if `until` is the one out of the valid range then `0 <= from <= length` is true,
which indicates we can return _slice_ from `from` to `length` with the length of `length - from`.

To satisfy the condition 4, 5 and 6, we need to check if `from <= until` is met.
To do this, ByteString1#slice(...) has the corresponding logic.