Use dropRight(...) instead of take(...) for the best use-case
We can expect that dropRight(M) is used to _drop_ elements from the right-hand side of the string
where M <= N/2 and N is the length of the string.
If this is the case, it can be said that (but not always the case as ByteString has Vector[ByteString1] data)
dropRight(M) is more efficient than take(N - M) because `M <= N - M` is satisfied.

Additionally, as _fullDrops_ represents the number of elements to _dropRight(N)_, some modifications was done :)