=htc #386 quote all parameter values in Content-Disposition header (#471)
Original RFC 2616 and newer RFC 6266 give different grammars / recommendations whether
to quote parameter values. It seems that current software is still confused about what to accept and what not.

Appendix A of RFC 6266 says:

      RFC 2616 only allows "quoted-string" for the filename parameter.
      This would be an exceptional parameter syntax, and also doesn't
      reflect actual use.

This doesn't mean that quotes are strictly required, so, given the reports of breakage
without quotes, let's force quotes for now for Content-Disposition parameters.