asOutputStream did never complete on close and exposed a unsafe callback #20503, #21864
StreamConverters.asOutputStream did never complete if the buffer size was greater than zero,
this lead to a bad state, where the close was unblocked, but the stage was never completed.
Also the OutputStream uses getAsyncCallback which had the problem that it always lead to races,
if flush or close was called to early.