Faster ByteString with extended Binary Indexed Tree
Summary:
By introducing additional data structure into ByteStrings[1], we can boost
the performance of ByteString operations, such as `take(n)`, `takeRight(n)`,
`drop(n)`, `dropRight(n)`, `apply(idx)` and `slice(start, until)`. As you can
find from the benchmark result, it's 10x to 50x faster than before (in the
end some operations are 100x to 300x times faster than that of baseline!).
One cons is we have some speed down of `++(append)` operation due to the
introduction of additional data(about 5 to 10% slowdown).

Problem:
As our ByteStrings contains multiple `ByteString1` as `Vector[ByteString1]`,
iteration of ByteString(vector) is not fast. As a result, some operation,
such as `take(N)` and `drop(N)`, is slow especially when the input size `N`
is big. The bigger the `N`, the more time it takes to finish the operation
because we need to keep searching for elements of the vector to find the
right place of `take` or `drop` for example.

Solution:
I was wondering if we knew the right place to `take` or `drop`. If we knew
the information, then there should be speed up of these operations.

The information we need is simply a nice _index_ table representing each
length of a vector(internal strings of ByteString).
What we want are:
1. fast searching for index of `take`, `drop` and so on.
2. minimum time consumption when appending `ByteStrings`, which we can't
   avoid as we introduce a new data structure (whose cost must be greater
   than 0).

To satisfy these conditions, I chose Binary Indexed Tree(BIT)[2] with some
extension.
The basic characteristics of BIT are:
1. easy to find sum between 1 to i-th elements, which is necessary to find
   index efficiently for our `take` operation for example.
2. easy to create the data structure using `Array[Int]` so we can expect
   a good performance. Actually, as the benchmark shows, `append` operation is
   not very _cheap_ so we need to discuss the best way to handle this issue.

More detailed explanation of our BIT is given in the source code.

[1] https://github.com/akka/akka/blob/v2.4.16/akka-actor/src/main/scala/akka/util/ByteString.scala#L401
[2] https://www.topcoder.com/community/data-science/data-science-tutorials/binary-indexed-trees/