Improve the server-side log message when default-host-header is not set
When the server receives a request it will try to reconstruct the
effective request URI as defined in HTTP/1.1 [1]. Requests using
HTTP/1.0 will use the same code path but since the `Host` header is not
required in HTTP/1.0 Akka HTTP requires that a default `Host` header
value is configured in `akka.http.server.default-host-header`. In
addition to the documentation in `reference.conf` make this more clear
by also hinting to the setting in the error log as discussed in #596.

 [1]: https://tools.ietf.org/html/rfc7230#section-5.5