=pro #222 reenable mima
The basic structure was copied over from akka. Now problems are not defined
in source any more but in separate files to be put into the `project/mima-filters`
directory. This should speed up building and updating the project.

It is still required to reload the project after such a file was added or
changed because the sbt key it provides is a `SettingKey` which is only
evaluated once.

Fixes #222.