=act #22568 reduce spurious ResumeReading log messages
ResumeReading will likely end up in DeadLetters when the connection is
torn down at the same time as the user tries to read more data from the
connection.

For the same reason ResumeReading may arrive in peerSentEOF state. We
ignore it here to get rid of "unhandled message from Actor ... : ResumeReading"
warnings.

This will remove the spurious warnings in these cases.