Use media type parameters in content negotiation (#963) (#964)
* Use media type parameters in content negotiation (#963)

https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.1 describes how
to interpret the Accept header during content negotiation.

Notably, it specifies "The media type quality factor associated with a given
type is determined by finding the media range with the highest precedence which
matches that type", and then gives some examples.

'matching' is implemented in akka-http in `MediaRange::matches`, but for
`MediaRange.One` and custom media ranges fails to take into account parameters.

'precedence' is implemented in akka-http when determining the sorting of
`acceptedMediaRanges` in `MediaTypeNegotiator`, but for non-wildcard media
types sorts only by the qValue ignoring the parameters.

We do want to take into account the `qValue` when determining the
order of `acceptedMediaRanges`, to make sure to select the right one when
encountering a media type that matches multiple ranges with the same
specificity but a different qValue. The specificity (number of parameters)
should have precedence though.