=htc #1013 make sure pool logs have set "PoolGateway" as logClass
Before, the log source was the descriptive string that describes
the pool but that can be hard to match on in logging backends. Now,
logging backends will see "PoolGateway" as the source for all pool
logging statements (still not quite the common practice but will
simplify to filter on all messages relating to the client pool).

Fixes #1013.