+htp #360 Allow Java-implemented ContentTypeResolver (#984)
The JavaMapping class assumes that all Java interfaces have
factory methods that only instantiate their Scala subclasses
from the scala DSL. That's not the case for ContentTypeResolver,
which is only a Java interface and a Scala trait. It's intended
to be implemented as a lambda / function.

However, when doing that from Java, the Scala trait is not
instantiated, upsetting the JavaMapping magic.

This commit changes ContentTypeResolver to no longer use the
JavaMapping conversion, but rather convert explicitly. By making
it a Java 8 functional interface, it can now be directly
implemented as a lambda. Binary compatibility is hopefully
unaffected, since the default method isn't slapped on until
actual Java implementations (which couldn't exist until now).

Fixes #360.