=htp #1072 move special non-2xx handling from RequestContextImpl to fromStatusCodeAndHeadersAndValue marshaller
This is an alternative solution for akka/akka#19397 that replaces the one
introduced in akka/akka#19849.

The former solution placed the special handling inside the main processing
path of requests in Akka HTTP. This solution is simpler and fixes the issue
in a narrower way.

The code comment explains the solution:

For non-2xx status, add an opaque fallback marshalling
using the first returned marshalling that will be used if the
result wouldn't otherwise be accepted. The reasoning is that users
often use routes like  `complete(400 -> "Illegal request in this context")`
to fail a route (instead of using rejection handling). If the client uses
narrow Accept headers like `Accept: application/json` the user
supplied error message will not be rendered because it will only
accept `text/plain` but not `application/json` responses and fail
the whole request with a "406 Not Acceptable" response.

Adding the opaque fallback rendering will give an escape hatch for
those situations.

Added more tests to `FromStatusCodeAndXYZMarshallerSpec` (formerly known
as `ContentNegotiationGivenResponseCodeSpec`) to cover the new behavior.