=pro don't exclude mima file changes from PR validation
The filter was added here before in akka/akka to make sure that mima
rules could be added without triggering PR validation for all subprojects.

With the external files this doesn't apply to akka-http any more. As a
side-effect we now rerun all tests even when changing Mima.scala. That
shouldn't be too much of a burden because Mima.scala changes are seldom
enough.