Fix #1134: Create representation of Content-Type with missing charset
The text/xml content-type (and possibly others) is best used with a
charset parameter, and akka inforces this on the type level. That's
great practice, but _incoming_ requests can certainly have a
Content-Type header of "text/xml" without a charset. Akka currently
just blindly adds "UTF-8", which is incorrect and a potential problem.

This commit adds a specific type to the ContentType hierarchy to represent
"a media type that should have had a charset but didn't".