+htt Support separation of route execution from checking in the Java DSL
Port the ScalatestRouteTestSpec to JUnitRouteTestTest and move await of
the route result from RouteTest to TestRouteResult similar to how the
Scala RouteTest works. This breaks binary compatibility of the testkit,
however, no guarantees are given for that part of the Akka HTTP API.

Preliminary work needed to implement the Java API for #39 (integration
route testing).