#1341 Return BadRequest when size of FormData exceeds limit of `withSizeLimit` directive
Actual behavior:
withSizeLimit in combination with formFieldMap directive responds with NotFound status code when a size of FormData exceeds the limit.

Expected behavior:
The combination of these directives should respond with BadRequest in a case of too large FormData, as it is implemented for too large Entity.

In order to fix this, handling of EntityStreamSizeException was added to FormFieldDirectives.