=htc #864 fix race condition in ResponseRendererSpec
The idea is to provide a stage / API that allows to find out the result
of running elements through a flow. The race condition previously was that
finding out whether the flow was completed could only be observed by waiting
for some amount of time.

This adds a new test utility stage that injects an event into the mailbox
of the ActorGraphInterpeter (via AsyncCallback) which is hoped to be only
processed after all internal events have been processed. This seems to work
reliably as long as the complete flow is fused and synchronous and the whole
processing needs less events than the event queue processes in one go.

Fixes #864