=htc lookup predefined header parsers as early as possible
Previously, every header parsing attempt had to go through the
`DynamicRuleDispatch`, a binary search to find the parsing rule for an
HTTP header name.

However, the `HttpHeaderParser` already has a parsing tree that with a node
for every header name.

This change divides header parsing into two steps: looking up the rule
and invoking the parser. Looking up the rule now only needs to be once
per node in the `HttpHeaderParser` and afterwards it can just run
without going through the dispatch.