Further stream shape cleanups (#23633)
* =str move non-boilerplaty shapes out of templates

This will improve code navigation for those classes.

 - FanInShape, FanInShape1N, UniformFanInShape
 - FanOutShape, UniformFanOutShape

* =str deprecate FanInShape1N

It was not tested, used or documented anywhere. With one input of type T0,
n inputs of type T1 and one output of type O, it is also arguably not
particularly generic to warrant its inclusion.

* !str remove unnecessary copyFromPorts implementation in Fan{In,Out}Shape

Strictly speaking this is a binary incompatible change. It would only affect
users that explicitly called `copyFromPorts` (but for what reason?).

* =str #23636 fix deprecation message not to use backticks

Backticks break genjavadoc / unidoc.

Fixes #23636.