Introduce 'PersistentActorContext' and ctx.persist()
I'm not sure yet whether I like it: while it does get rid of the
extra generic parameters when using `andThen`, it is kind of confusing
to expose both `Persist()` and `ctx.persist()`, and I'm not a big fan of
the latter because it looks so much like a side-effect even though it isn't.

I wonder if modeling the `andThen` as an `Effect` and having a function to
concatenate effects would turn out nicer.