Improve error message when pushing unpulled outlet
If an outlet is pushed to before it has been pulled, the error message
is "Cannot push twice", which makes sense if after receiving a pull, you
push twice, but makes no sense if you never received a pull in the first
place. This improves the error message to avoid confusion.