=doc Fix #1541: Bump number of unique requests to 500 (#1544)
The create cache tests demonstrate that if we set a small max capacity
(say 50 entries) and make more than that amount of unique requests and
finally redo the initial request then a new cache entry should be
generated. That is the entry of the initial request has been evicted and
regenerated.

The problem is that the LFU-cache implementation does not guarantee when
eviction happens. It might happen at next write time or as we clearly
see in #1541 at some later convenience. As a workaround bump the number
of requests to be much higher.