+htc #1312 new host connection pool implementation
A new implementation of the HostConnectionPool. The basic idea is to replace
the former complicated streaming pipeline of `PoolFlow`, `PoolConductor`,
and `PoolSlot` into a single stage that handles all aspects to get rid
of all the small race condition issues that exist in the current ("legacy")
pool implementation.

The new pool implementation is split into two basic classes

 * akka.http.impl.engine.client.pool.NewHostConnectionPool that provides all
   the infrastructure and event handling to drive the pool
 * akka.http.impl.engine.client.pool.SlotState that contains only the logic
   to handle state changes of a single pool slot