=htc try to keep pool running when a single slot fails
This is a best effort try not having to kill a pool if a single slot gets
into a weird state. In the best case, that will allow the pool to keep going
in the presence of bugs that are only triggered rarely. On the other hand,
the risk is that we keep a pool alive that might be in a weird state leading
to more problem further down the road.